Refine module organization based on PR review feedback
Address PR #271 review comments to improve symbol organization:

vl-convert-rs/src/converter/:
- Extract value_or_string.rs from config.rs (ValueOrString enum + impls)
- Extract permissions.rs from config.rs (domain matching, build_permissions)
- Move GoogleFontRequest from config.rs to fonts.rs
- Extract plugin.rs from worker_pool.rs (plugin bundling/resolution)
- Split inner.rs (2524 lines) into inner/ module directory with 6 focused
  submodules: mod.rs, runtime.rs, conversions.rs, rendering.rs, fonts.rs,
  init.rs
- Move VlConverterInner struct from mod.rs into inner/ module

vl-convert-python/src/:
- Rename helpers.rs to utils.rs
- Move font helpers (parse_google_fonts_arg, effective_google_fonts) to
  fonts.rs
- Move config helpers (ConverterConfigOverrides, converter_read_handle,
  load_config_inner, etc.) to config.rs

vl-convert/src/:
- Rename io_helpers.rs to io_utils.rs
