+++ vl-convert/src/handlers.rs
@@ -4,7 +4,7 @@ use vl_convert_rs::converter::{
 };
 use vl_convert_rs::{anyhow, anyhow::bail};
 
-use crate::io_helpers::{
+use crate::io_utils::{
     parse_as_json, parse_format_locale_option, parse_time_format_locale_option, parse_vl_version,
     read_config_json, read_input_string, write_output_binary, write_output_string,
 };
